+++ src/AnnotatedRoutes/tests/App/App.php
@@ -5,15 +5,16 @@ declare(strict_types=1);
 namespace Spiral\Tests\Router\App;
 
 use Spiral\Console\Console;
+use Spiral\Core\Container;
 use Spiral\Framework\Kernel;
-use Spiral\Http\Bootloader\DiactorosBootloader;
 use Spiral\Http\Http;
+use Spiral\Nyholm\Bootloader\NyholmBootloader;
 use Spiral\Router\Bootloader\AnnotatedRoutesBootloader;
 
 class App extends Kernel
 {
     protected const LOAD = [
-        DiactorosBootloader::class,
+        NyholmBootloader::class,
         AnnotatedRoutesBootloader::class,
     ];
 
@@ -25,5 +26,10 @@ class App extends Kernel
     public function getConsole(): Console
     {
         return $this->container->get(Console::class);
+    }
+
+    public function getContainer(): Container
+    {
+        return $this->container;
     }
 }