+++ tests/app/src/TestApp.php
@@ -22,6 +22,7 @@ use Spiral\Bootloader\ExceptionHandlerBootloader;
 use Spiral\Console\Console;
 use Spiral\Core\Container;
 use Spiral\Framework\Kernel;
+use Spiral\Nyholm\Bootloader\NyholmBootloader;
 use Spiral\Stempler\Bootloader\StemplerBootloader;
 
 class TestApp extends Kernel
@@ -41,7 +42,7 @@ class TestApp extends Kernel
         \Spiral\Console\Bootloader\ConsoleBootloader::class,
 
         // HTTP extensions
-        \Spiral\Http\Bootloader\DiactorosBootloader::class,
+        NyholmBootloader::class,
         Bootloader\Http\RouterBootloader::class,
         Bootloader\Http\ErrorHandlerBootloader::class,
         Bootloader\Http\JsonPayloadsBootloader::class,